requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
timm==0.9.0
lightning==2.1.3
Pillow==10.0.0
requests==2.31.0
pandas==2.2.0
numpy==1.26.3
python-dotenv>=1.0.1